Rename bootstrap file to bootstrap3
So we have a clean naming to separate it from the version 4.
