+++ lib/generators/simple_form/templates/config/initializers/simple_form_bootstrap3.rb
